# openapi/openapi.yaml
openapi: 3.1.2
info:
  title: LibOps API
  version: 1.0.0
  description: LibOps control plane API to run docker compose projects on Google Cloud
  contact:
    name: LibOps
    url: https://github.com/libops/api
  license:
    name: MIT
    url: https://github.com/libops/api/blob/main/LICENSE
servers:
- url: https://api.libops.io
  description: Production server
paths:
...
  /libops.v1.AccountService/GetAccountByEmail:
    get:
      tags:
      - libops.v1.AccountService
      summary: Get account information by email (for Terraform provider lookups)
      description: Get account information by email (for Terraform provider lookups)
      operationId: libops.v1.AccountService.GetAccountByEmail.get
      parameters:
      - name: Connect-Protocol-Version
        in: header
        required: true
        schema:
          $ref: '#/components/schemas/connect-protocol-version'
      - name: Connect-Timeout-Ms
        in: header
        schema:
          $ref: '#/components/schemas/connect-timeout-header'
      - name: message
        in: query
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/libops.v1.GetAccountByEmailRequest'
      - name: encoding
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/encoding'
      - name: base64
        in: query
        schema:
          $ref: '#/components/schemas/base64'
      - name: compression
        in: query
        schema:
          $ref: '#/components/schemas/compression'
      - name: connect
        in: query
        schema:
          $ref: '#/components/schemas/connect'
      responses:
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/connect.error'
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/libops.v1.GetAccountByEmailResponse'
    post:
      tags:
      - libops.v1.AccountService
      summary: Get account information by email (for Terraform provider lookups)
      description: Get account information by email (for Terraform provider lookups)
      operationId: libops.v1.AccountService.GetAccountByEmail
      parameters:
      - name: Connect-Protocol-Version
        in: header
        required: true
        schema:
          $ref: '#/components/schemas/connect-protocol-version'
      - name: Connect-Timeout-Ms
        in: header
        schema:
          $ref: '#/components/schemas/connect-timeout-header'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/libops.v1.GetAccountByEmailRequest'
        required: true
      responses:
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/connect.error'
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/libops.v1.GetAccountByEmailResponse'
...
  /libops.v1.AdminAccountService/CreateAccount:
    post:
      tags:
      - libops.v1.AdminAccountService
      summary: Create a new account
      description: Create a new account
      operationId: libops.v1.AdminAccountService.CreateAccount
      parameters:
      - name: Connect-Protocol-Version
        in: header
        required: true
        schema:
          $ref: '#/components/schemas/connect-protocol-version'
      - name: Connect-Timeout-Ms
        in: header
        schema:
          $ref: '#/components/schemas/connect-timeout-header'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/libops.v1.CreateAccountRequest'
        required: true
      responses:
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/connect.error'
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/libops.v1.CreateAccountResponse'
...
  /libops.v1.AdminAccountService/GetAccountByEmail:
    get:
      tags:
      - libops.v1.AdminAccountService
      summary: Get account information by email
      description: Get account information by email
      operationId: libops.v1.AdminAccountService.GetAccountByEmail.get
      parameters:
      - name: Connect-Protocol-Version
        in: header
        required: true
        schema:
          $ref: '#/components/schemas/connect-protocol-version'
      - name: Connect-Timeout-Ms
        in: header
        schema:
          $ref: '#/components/schemas/connect-timeout-header'
      - name: message
        in: query
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/libops.v1.AdminGetAccountByEmailRequest'
      - name: encoding
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/encoding'
      - name: base64
        in: query
        schema:
          $ref: '#/components/schemas/base64'
      - name: compression
        in: query
        schema:
          $ref: '#/components/schemas/compression'
      - name: connect
        in: query
        schema:
          $ref: '#/components/schemas/connect'
      responses:
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/connect.error'
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/libops.v1.AdminGetAccountByEmailResponse'
    post:
      tags:
      - libops.v1.AdminAccountService
      summary: Get account information by email
      description: Get account information by email
      operationId: libops.v1.AdminAccountService.GetAccountByEmail
      parameters:
      - name: Connect-Protocol-Version
        in: header
        required: true
        schema:
          $ref: '#/components/schemas/connect-protocol-version'
      - name: Connect-Timeout-Ms
        in: header
        schema:
          $ref: '#/components/schemas/connect-timeout-header'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/libops.v1.AdminGetAccountByEmailRequest'
        required: true
      responses:
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/connect.error'
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/libops.v1.AdminGetAccountByEmailResponse'
...
  /libops.v1.AdminAccountService/UpdateAccount:
    post:
      tags:
      - libops.v1.AdminAccountService
      summary: Update account information
      description: Update account information
      operationId: libops.v1.AdminAccountService.UpdateAccount
      parameters:
      - name: Connect-Protocol-Version
        in: header
        required: true
        schema:
          $ref: '#/components/schemas/connect-protocol-version'
      - name: Connect-Timeout-Ms
        in: header
        schema:
          $ref: '#/components/schemas/connect-timeout-header'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/libops.v1.UpdateAccountRequest'
        required: true
      responses:
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/connect.error'
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/libops.v1.UpdateAccountResponse'
...
components:
  schemas:
    base64:
      type: boolean
      title: base64
      description: Specifies if the message query param is base64 encoded, which may
        be required for binary data
    compression:
      title: compression
      enum:
      - identity
      - gzip
      - br
      description: Which compression algorithm to use for this request
    connect:
      title: connect
      enum:
      - v1
      description: Define the version of the Connect protocol
    connect-protocol-version:
      type: number
      title: Connect-Protocol-Version
      enum:
      - 1
      description: Define the version of the Connect protocol
      const: 1
    connect-timeout-header:
      type: number
      title: Connect-Timeout-Ms
      description: Define the timeout, in ms
    connect.error:
      type: object
      properties:
        code:
          type: string
          examples:
          - not_found
          enum:
          - canceled
          - unknown
          - invalid_argument
          - deadline_exceeded
          - not_found
          - already_exists
          - permission_denied
          - resource_exhausted
          - failed_precondition
          - aborted
          - out_of_range
          - unimplemented
          - internal
          - unavailable
          - data_loss
          - unauthenticated
          description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
        message:
          type: string
          description: A developer-facing error message, which should be in English.
            Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details]
            field, or localized by the client.
        details:
          type: array
          items:
            $ref: '#/components/schemas/connect.error_details.Any'
          description: A list of messages that carry the error details. There is no
            limit on the number of messages.
      title: Connect Error
      additionalProperties: true
      description: 'Error type returned by Connect: https://connectrpc.com/docs/go/errors/#http-representation'
    connect.error_details.Any:
      type: object
      properties:
        type:
          type: string
          description: 'A URL that acts as a globally unique identifier for the type
            of the serialized message. For example: `type.googleapis.com/google.rpc.ErrorInfo`.
            This is used to determine the schema of the data in the `value` field
            and is the discriminator for the `debug` field.'
        value:
          type: string
          format: binary
          description: The Protobuf message, serialized as bytes and base64-encoded.
            The specific message type is identified by the `type` field.
        debug:
          oneOf:
          - type: object
            title: Any
            additionalProperties: true
            description: Detailed error information.
          discriminator:
            propertyName: type
          title: Debug
          description: Deserialized error detail payload. The 'type' field indicates
            the schema. This field is for easier debugging and should not be relied
            upon for application logic.
      additionalProperties: true
      description: Contains an arbitrary serialized message along with a @type that
        describes the type of the serialized message, with an additional debug field
        for ConnectRPC error details.
    encoding:
      title: encoding
      enum:
      - proto
      - json
      description: Define which encoding or 'Message-Codec' to use
...
    libops.v1.Account:
      type: object
      properties:
        accountId:
          type: string
          title: account_id
          description: Unique account identifier (UUID)
        email:
          type: string
          title: email
          description: User's email address
        name:
          type: string
          title: name
          description: Display name
        githubUsername:
          type: string
          title: github_username
          description: GitHub username for integration
          nullable: true
        role:
          title: role
          description: System-wide role
          $ref: '#/components/schemas/libops.v1.AccountRole'
        status:
          title: status
          description: Account status
          $ref: '#/components/schemas/libops.v1.AccountStatus'
        authMethod:
          title: auth_method
          description: How the user authenticates
          $ref: '#/components/schemas/libops.v1.common.AuthMethod'
        verified:
          type: boolean
          title: verified
          description: Email verification status
        verifiedAt:
          type: string
          title: verified_at
          description: Timestamp when verified (RFC3339)
          nullable: true
      title: Account
      additionalProperties: false
    libops.v1.AccountRole:
      type: string
      title: AccountRole
      enum:
      - ACCOUNT_ROLE_UNSPECIFIED
      - ACCOUNT_ROLE_USER
      - ACCOUNT_ROLE_ADMIN
    libops.v1.AccountStatus:
      type: string
      title: AccountStatus
      enum:
      - ACCOUNT_STATUS_UNSPECIFIED
      - ACCOUNT_STATUS_ACTIVE
      - ACCOUNT_STATUS_SUSPENDED
      - ACCOUNT_STATUS_DELETED
...
    libops.v1.AdminGetAccountByEmailRequest:
      type: object
      properties:
        email:
          type: string
          title: email
      title: AdminGetAccountByEmailRequest
      additionalProperties: false
    libops.v1.AdminGetAccountByEmailResponse:
      type: object
      properties:
        account:
          title: account
          $ref: '#/components/schemas/libops.v1.Account'
      title: AdminGetAccountByEmailResponse
      additionalProperties: false
...
    libops.v1.CreateAccountRequest:
      type: object
      properties:
        email:
          type: string
          title: email
        name:
          type: string
          title: name
        githubUsername:
          type: string
          title: github_username
          nullable: true
      title: CreateAccountRequest
      additionalProperties: false
    libops.v1.CreateAccountResponse:
      type: object
      properties:
        account:
          title: account
          $ref: '#/components/schemas/libops.v1.Account'
      title: CreateAccountResponse
      additionalProperties: false
...
    libops.v1.GetAccountByEmailRequest:
      type: object
      properties:
        email:
          type: string
          title: email
      title: GetAccountByEmailRequest
      additionalProperties: false
    libops.v1.GetAccountByEmailResponse:
      type: object
      properties:
        account:
          title: account
          $ref: '#/components/schemas/libops.v1.OrganizationAccount'
      title: GetAccountByEmailResponse
      additionalProperties: false
...
    libops.v1.OrganizationAccount:
      type: object
      properties:
        accountId:
          type: string
          title: account_id
          description: Unique account identifier (UUID)
        email:
          type: string
          title: email
          description: User's email address
        name:
          type: string
          title: name
          description: Display name
        authMethod:
          title: auth_method
          description: How the user authenticates
          $ref: '#/components/schemas/libops.v1.common.AuthMethod'
        verified:
          type: boolean
          title: verified
          description: Email verification status
      title: OrganizationAccount
      additionalProperties: false
...
    libops.v1.UpdateAccountRequest:
      type: object
      properties:
        accountId:
          type: string
          title: account_id
        name:
          type: string
          title: name
          nullable: true
        githubUsername:
          type: string
          title: github_username
          nullable: true
        updateMask:
          title: update_mask
          $ref: '#/components/schemas/google.protobuf.FieldMask'
      title: UpdateAccountRequest
      additionalProperties: false
    libops.v1.UpdateAccountResponse:
      type: object
      properties:
        account:
          title: account
          $ref: '#/components/schemas/libops.v1.Account'
      title: UpdateAccountResponse
      additionalProperties: false
...
    libops.v1.common.AuthMethod:
      type: string
      title: AuthMethod
      enum:
      - AUTH_METHOD_UNSPECIFIED
      - AUTH_METHOD_GOOGLE
      - AUTH_METHOD_USERPASS
      - AUTH_METHOD_GCLOUD
      description: AuthMethod represents how a user authenticates to libops
...
  securitySchemes:
    oauth2:
      type: oauth2
      description: OAuth 2.0 authentication via Vault OIDC
      flows:
        authorizationCode:
          authorizationUrl: /auth/oauth/authorize
          tokenUrl: /auth/oauth/token
          scopes:
            read:user: Read user account information
            write:user: Update user account information
            read:organizations: Read user's organizations
            read:organization: Read organization details
            write:organization: Update organization
            delete:organization: Delete organization
            read:projects: Read organization projects
            create:projects: Create organization projects
            write:projects: Update organization projects
            delete:projects: Delete organization projects
            read:members: Read organization/project/site members
            write:members: Manage organization/project/site members
            delete:members: Remove organization/project/site members
            read:invoices: Read organization invoices
            read:firewall: Read firewall rules
            write:firewall: Update firewall rules
            delete:firewall: Delete firewall rules
            read:site: Read site details
            write:site: Update site configuration
            delete:site: Delete site
            read:sites: Read project sites
            write:sites: Update project sites
            delete:sites: Delete project sites
            promote_sites: Promote sites between environments
            read:project: Read project details
            write:project: Update project
            delete:project: Delete project
            manage_secrets: Read, write, and delete secrets
            admin:system: Full system administrative access
    apiKey:
      type: http
      scheme: bearer
...
security: []
...